requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
passlib
bcrypt
Faker
starlette==0.17.1
black
pylint
python-multipart
python-jose[cryptography]